- Now PdfDocumentData acts as an independent class and actions are performed directly with it. - Now it is validated the type of output to be file or directory according to the option used.
